refactor: InputView파일에서 BridgeGameSetter 분리
